SL-ObjectDetection-Faster-R-CNN/requirements.txt
# Core ML libraries
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# Computer vision
opencv-python>=4.5.0
Pillow>=8.0.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
tensorboard>=2.6.0

# Utilities
tqdm>=4.62.0
PyYAML>=5.4.0
albumentations>=1.0.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Development
black>=21.0.0
isort>=5.9.0
flake8>=3.9.0
pre-commit>=2.15.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
